packages/ml/src/ml/models/embeddings/mlflow_embed/pyproject.toml
[project]
name = "mlflow-embed"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10,<=3.13"
dependencies = [
    "python-dotenv>=1.1.0",
 "mlflow>=2.22.0",
 "go-task-bin>=3.43.3",
]
